fix(bench): move memory benchmark harness into daemon
