+++ crates/daemon/Cargo.toml
@@ -44,6 +44,7 @@ loongclaw-spec = { path = "../spec" }
 serde.workspace = true
 serde_json.workspace = true
 reqwest.workspace = true
+rusqlite.workspace = true
 tokio.workspace = true
 time.workspace = true
 toml.workspace = true
